chore: refactor from dataset manager to dataset item repository CRUD methods
